Installed helm, Added Specific K8's version and release.

+++ Vagrantfile
@@ -7,6 +7,8 @@ servers = [
         :type => "master",
         :box => "ubuntu/xenial64",
         :box_version => "20180831.0.0",
+        :k8s_version => "1.15.0-00",
+        :k8s_release => "stable-1.15",
         :eth1 => "192.168.205.10",
         :mem => "2048",
         :cpu => "2"
@@ -16,6 +18,8 @@ servers = [
         :type => "node",
         :box => "ubuntu/xenial64",
         :box_version => "20180831.0.0",
+        :k8s_version => "1.15.0-00",
+        :k8s_release => "stable-1.15",
         :eth1 => "192.168.205.11",
         :mem => "2048",
         :cpu => "2"
@@ -25,6 +29,8 @@ servers = [
         :type => "node",
         :box => "ubuntu/xenial64",
         :box_version => "20180831.0.0",
+        :k8s_version => "1.15.0-00",
+        :k8s_release => "stable-1.15",
         :eth1 => "192.168.205.12",
         :mem => "2048",
         :cpu => "2"
@@ -36,13 +42,16 @@ $configureBox = <<-SCRIPT
 
     # install docker v17.03
     # reason for not using docker provision is that it always installs latest version of the docker, but kubeadm requires 17.03 or older
+
+    k8s_version=$1
+    cd $HOME
     apt-get update
     apt-get install -y apt-transport-https ca-certificates curl software-properties-common
     curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add -
     add-apt-repository "deb https://download.docker.com/linux/$(. /etc/os-release; echo "$ID") $(lsb_release -cs) stable"
     apt-get update && apt-get install -y docker-ce=$(apt-cache madison docker-ce | grep 17.03 | head -1 | awk '{print $3}')
 
-    # run docker commands as vagrant user (sudo not required)
+    # run docker commands as root user (sudo not required)
     usermod -aG docker vagrant
 
     # install kubeadm
@@ -52,7 +61,7 @@ $configureBox = <<-SCRIPT
     deb http://apt.kubernetes.io/ kubernetes-xenial main
 EOF
     apt-get update
-    apt-get install -y kubelet kubeadm kubectl
+    apt-get install -y kubelet=${k8s_version} kubeadm=${k8s_version} kubectl=${k8s_version}
     apt-mark hold kubelet kubeadm kubectl
 
     # kubelet requires swap off
@@ -64,9 +73,9 @@ EOF
     # ip of this box
     IP_ADDR=`ifconfig enp0s8 | grep Mask | awk '{print $2}'| cut -f2 -d:`
     # set node-ip
-    sudo wget https://raw.githubusercontent.com/elanadmin/kubernetes-cluster/master/config/kubelet -O /etc/default/kubelet
+    cp /vagrant/config/kubelet /etc/default/kubelet
     if [ -f "/bin/systemd" ];then
-      sudo wget https://raw.githubusercontent.com/elanadmin/kubernetes-cluster/master/config/daemon.json -O /etc/docker/daemon.json
+      cp /vagrant/config/daemon.json /etc/docker/daemon.json
       sudo systemctl restart docker
     fi
     sudo sed -i "/^[^#]*KUBELET_EXTRA_ARGS=/c\KUBELET_EXTRA_ARGS=--node-ip=$IP_ADDR" /etc/default/kubelet
@@ -85,29 +94,51 @@ $configureMaster = <<-SCRIPT
     # ip of this box
     IP_ADDR=`ifconfig enp0s8 | grep Mask | awk '{print $2}'| cut -f2 -d:`
 
-    # install k8s master
-    HOST_NAME=$(hostname -s)
-    kubeadm init --apiserver-advertise-address=$IP_ADDR --apiserver-cert-extra-sans=$IP_ADDR  --node-name $HOST_NAME --pod-network-cidr=172.16.0.0/16
+    k8s_release=$1
+    cd $HOME
+    if [ ! -f /etc/kubernetes/admin.conf ];then
+      # install k8s master
+      HOST_NAME=$(hostname -s)
+      kubeadm init --kubernetes-version ${k8s_release} --apiserver-advertise-address=$IP_ADDR --apiserver-cert-extra-sans=$IP_ADDR  --node-name $HOST_NAME --pod-network-cidr=172.16.0.0/16
 
-    #copying credentials to regular user - vagrant
-    sudo --user=vagrant mkdir -p /home/vagrant/.kube
-    cp -i /etc/kubernetes/admin.conf /home/vagrant/.kube/config
-    chown $(id -u vagrant):$(id -g vagrant) /home/vagrant/.kube/config
+      #copying credentials 
+      mkdir -p $HOME/.kube
+      cp -i /etc/kubernetes/admin.conf $HOME/.kube/config
 
-    # install Calico pod network addon
-    export KUBECONFIG=/etc/kubernetes/admin.conf
-    kubectl apply -f https://raw.githubusercontent.com/elanadmin/kubernetes-cluster/master/calico/calico.yaml
+      # install Calico pod network addon
+      export KUBECONFIG=/etc/kubernetes/admin.conf
+      kubectl apply -f https://raw.githubusercontent.com/elanadmin/kubernetes-cluster/master/calico/calico.yaml
 
-    kubeadm token create --print-join-command >> /etc/kubeadm_join_cmd.sh
-    chmod +x /etc/kubeadm_join_cmd.sh
+      kubeadm token create --print-join-command >> /etc/kubeadm_join_cmd.sh
+      chmod +x /etc/kubeadm_join_cmd.sh
+      kubectl taint nodes --all node-role.kubernetes.io/master-
+    fi
+
+    # Install Helm Package Manager
+    if [ ! -f /usr/sbin/helm ];then
+      wget https://kubernetes-helm.storage.googleapis.com/helm-v2.8.2-linux-amd64.tar.gz
+      wget https://git.io/get_helm.sh -O get_helm.sh
+      tar -zxvf helm-v2.8.2-linux-amd64.tar.gz
+      sudo cp -rp linux-amd64/helm /usr/sbin/
+      echo export HELM_HOME="$HOME/.helm" >> $HOME/.bash_profile
+      export HELM_HOME="$HOME/.helm"
+      helm init | tee -a ~/helmet.init
+      kubectl create serviceaccount --namespace kube-system tiller 2> /dev/null
+      kubectl create clusterrolebinding tiller-cluster-rule --clusterrole=cluster-admin --serviceaccount=kube-system:tiller 2>/dev/null
+      kubectl patch deploy --namespace kube-system tiller-deploy -p '{"spec":{"template":{"spec":{"serviceAccount":"tiller"}}}}' 2>/dev/null
+    fi
 
 SCRIPT
 
 $configureNode = <<-SCRIPT
-    echo -e "\nThis is worker:\n"
-    apt-get install -y sshpass
-    sshpass -p "vagrant" scp -o StrictHostKeyChecking=no vagrant@192.168.205.10:/etc/kubeadm_join_cmd.sh .
-    sh ./kubeadm_join_cmd.sh
+    PWD=$(pwd)
+    cd $HOME
+    if [ ! -f $PWD/kubeadm_join_cmd.sh ];then
+      echo -e "\nThis is worker:\n"
+      apt-get install -y sshpass
+      sshpass -p "vagrant" scp -o StrictHostKeyChecking=no vagrant@192.168.205.10:/etc/kubeadm_join_cmd.sh .
+      sh ./kubeadm_join_cmd.sh
+    fi
 SCRIPT
 
 Vagrant.configure("2") do |config|
@@ -133,12 +164,17 @@ Vagrant.configure("2") do |config|
             # we cannot use this because we can't install the docker version we want - https://github.com/hashicorp/vagrant/issues/4871
             # config.vm.provision "docker"
 
-            config.vm.provision "shell", inline: $configureBox
+            k8s_version = opts[:k8s_version]
+            k8s_release = opts[:k8s_release]
+
+            config.vm.provision "shell", inline: $configureBox, :args => [k8s_version]
 
             if opts[:type] == "master"
-                config.vm.provision "shell", inline: $configureMaster
+              config.vm.provision "shell", inline: $configureMaster, :args => [k8s_release],
+                  run: "always"
             else
-                config.vm.provision "shell", inline: $configureNode
+                config.vm.provision "shell", inline: $configureNode,
+                  run: "always"
             end
 
         end
